9.4 - Custom Hooks
In this new class I learned how to use and type a new hook.
For example, using useRef to hold new values without the necessity for re-render this value. And the "!" after current to make sure this value isn't null.

So, the custom hook is placed in the regular TSX component in order to call the logic inside the created hook.
Just like in JS. However, we need to "import" the "default" "function" using the TS name, in this case "useInterval" and then, type the function.